fixed the way number of arguments was checked to be consistent across all commands

+++ handler_reset.go
@@ -6,6 +6,10 @@ import (
 )
 
 func handlerReset(s *state, cmd command) error {
+	if len(cmd.args) != 0 {
+		return fmt.Errorf("no arguments required")
+	}
+
 	err := s.db.DeleteAllUsers(context.Background())
 	if err != nil {
 		return fmt.Errorf("error reseting database: %v", err)
